requirements.txt
altgraph==0.17.3
importlib-metadata==6.6.0
jaraco.classes==3.2.3
keyring==23.13.1
macholib==1.16.2
more-itertools==9.1.0
pathspec==0.11.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyYAML==6.0
yamllint==1.31.0
zipp==3.15.0
